backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pandas>=2.2.0
statsforecast>=1.7.0
groq>=0.12.0
supabase>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
openpyxl>=3.1.0
slowapi>=0.1.9
python-jose[cryptography]>=3.3.0
langfuse>=2.0.0
pytest>=8.0.0
httpx>=0.27.0
pytest-asyncio>=0.23.0
